Products section refactor
