+++ components/molecules/index.ts
@@ -4,5 +4,18 @@ import UserNav from "./UserNav/UserNav";
 import LoginWithEmail from "./LoginWithEmail/LoginWithEmail";
 import CreateAccount from "./CreateAccount/CreateAccount";
 import ResetPassword from "./ResetPassword/ResetPassword";
+import GalleryComponent from "./Gallery/Gallery";
+import ImagesSection from "./ImagesSection/ImagesSection";
+import ProductDescription from "./ProductDescription/ProductDescription";
 
-export { InfoBar, MainNavigation, UserNav, LoginWithEmail, CreateAccount, ResetPassword };
+export {
+  InfoBar,
+  MainNavigation,
+  UserNav,
+  LoginWithEmail,
+  CreateAccount,
+  ResetPassword,
+  GalleryComponent,
+  ImagesSection,
+  ProductDescription,
+};
