+++ components/templates/Layout/Layout.styles.ts
@@ -3,7 +3,7 @@ import styled from "styled-components";
 export const StyledWrapper = styled.div`
   position: relative;
   display: grid;
-  grid-template-rows: auto auto 10rem 1fr auto;
+  grid-template-rows: auto auto auto 1fr auto;
   min-height: 100vh;
 `;
 
@@ -11,4 +11,4 @@ export const WidthContainer = styled.div`
   width: 100%;
   max-width: var(--maxWidth);
   margin: 0 auto;
-`
+`;
